Move `SettingSelect` component to root components dir to reuse it
