+++ src/components/SettingSelect/index.js
@@ -1,0 +1,3 @@
+import SettingSelect from './SettingSelect'
+
+export default SettingSelect
